fix get pokemon use case test
